+++ tests/testthat/test-acep_gemini.R
@@ -1,0 +1,8 @@
+
+test_that("ACEP gemini", {
+  skip_if_offline()
+  skip_on_cran()
+  texto <- 1:10
+  dimensiones <- length(texto)
+  expect_equal(dimensiones, length(texto))
+})